Change wctype_t from int to long.
On LP64 this brings us on par with the other libcs where wctype_t is 8
bytes.

Bug: 12875898

+++ libc/include/wchar.h
@@ -64,7 +64,7 @@ enum {
     WC_TYPE_MAX
 };
 
-typedef int wctype_t;
+typedef long wctype_t;
 
 #define  WEOF        ((wint_t)(-1))
